+++ asyncprawcore/requestor.py
@@ -23,6 +23,7 @@ class Requestor(object):
         reddit_url="https://www.reddit.com",
         session=None,
         loop=None,
+        timeout=TIMEOUT,
     ):
         """Create an instance of the Requestor class.
 
@@ -34,6 +35,8 @@ class Requestor(object):
             (Default: https://www.reddit.com)
         :param session: (Optional) A session to handle requests, compatible with
             aiohttp.ClientSession(). (Default: None)
+        :param timeout: (Optional) How many seconds to wait for the server to send data
+            before giving up. (Default: const.TIMEOUT)
 
         """
         if user_agent is None or len(user_agent) < 7:
@@ -49,14 +52,17 @@ class Requestor(object):
 
         self.oauth_url = oauth_url
         self.reddit_url = reddit_url
+        self.timeout = timeout
 
     async def close(self):
         """Call close on the underlying session."""
         return await self._http.close()
 
-    async def request(self, *args, timeout=TIMEOUT, **kwargs):
+    async def request(self, *args, timeout=None, **kwargs):
         """Issue the HTTP request capturing any errors that may occur."""
         try:
-            return await self._http.request(*args, timeout=timeout, **kwargs)
+            return await self._http.request(
+                *args, timeout=timeout or self.timeout, **kwargs
+            )
         except Exception as exc:
             raise RequestException(exc, args, kwargs)